+++ test/sagemaker_tests/tensorflow/inference/test/integration/local/test_multi_model_endpoint.py
@@ -11,6 +11,7 @@
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
+import encodings
 import json
 import os
 import subprocess
@@ -20,10 +21,10 @@ import time
 import pytest
 import requests
 
-from multi_model_endpoint_test_utils import make_invocation_request, make_list_model_request, \
-    make_get_model_request, make_load_model_request, make_unload_model_request
-
 PING_URL = 'http://localhost:8080/ping'
+INVOCATION_URL = 'http://localhost:8080/models/{}/invoke'
+MODELS_URL = 'http://localhost:8080/models'
+DELETE_MODEL_URL = 'http://localhost:8080/models/{}'
 
 
 @pytest.fixture(scope='session', autouse=True)
@@ -69,26 +70,41 @@ def container(request, docker_base_name, tag, runtime_config):
         subprocess.check_call('docker rm -f sagemaker-tensorflow-serving-test'.split())
 
 
+def make_invocation_request(data, model_name, content_type='application/json'):
+    headers = {
+        'Content-Type': content_type,
+        'X-Amzn-SageMaker-Custom-Attributes': 'tfs-method=predict'
+    }
+    response = requests.post(INVOCATION_URL.format(model_name), data=data, headers=headers)
+    return response.status_code, json.loads(response.content.decode(encodings.utf_8.getregentry().name))
+
+
+def make_list_model_request():
+    response = requests.get(MODELS_URL)
+    return response.status_code, json.loads(response.content.decode(encodings.utf_8.getregentry().name))
+
+
+def make_get_model_request(model_name):
+    response = requests.get(MODELS_URL + '/{}'.format(model_name))
+    return response.status_code, json.loads(response.content.decode(encodings.utf_8.getregentry().name))
+
+
+def make_load_model_request(data, content_type='application/json'):
+    headers = {
+        'Content-Type': content_type
+    }
+    response = requests.post(MODELS_URL, data=data, headers=headers)
+    return response.status_code, response.content.decode(encodings.utf_8.getregentry().name)
+
+
+def make_unload_model_request(model_name):
+    response = requests.delete(DELETE_MODEL_URL.format(model_name))
+    return response.status_code, response.content.decode(encodings.utf_8.getregentry().name)
+
+
 def test_ping():
     res = requests.get(PING_URL)
     assert res.status_code == 200
-
-
-def test_container_start_invocation_fail():
-    x = {
-        'instances': [1.0, 2.0, 5.0]
-    }
-    code, y = make_invocation_request(json.dumps(x), 'half_plus_three')
-    y = json.loads(y)
-    assert code == 404
-    assert "Model half_plus_three is not loaded yet." in str(y)
-
-
-def test_list_models_empty():
-    code, res = make_list_model_request()
-    res = json.loads(res)
-    assert code == 200
-    assert len(res) == 0
 
 
 def test_delete_unloaded_model():
@@ -96,7 +112,7 @@ def test_delete_unloaded_model():
     model_name = 'non-existing-model'
     code, res = make_unload_model_request(model_name)
     assert code == 404
-    assert 'Model {} is not loaded yet'.format(model_name) in res
+    assert res == '{} not loaded yet.'.format(model_name)
 
 
 def test_delete_model():
@@ -107,22 +123,35 @@ def test_delete_model():
     }
     code, res = make_load_model_request(json.dumps(model_data))
     assert code == 200
-    assert 'Successfully loaded model {}'.format(model_name) in res
+    assert res == 'Successfully loaded model {}'.format(model_name)
 
     x = {
         'instances': [1.0, 2.0, 5.0]
     }
     _, y = make_invocation_request(json.dumps(x), model_name)
-    y = json.loads(y)
     assert y == {'predictions': [3.5, 4.0, 5.5]}
 
     code_unload, res2 = make_unload_model_request(model_name)
     assert code_unload == 200
 
     code_invoke, y2 = make_invocation_request(json.dumps(x), model_name)
-    y2 = json.loads(y2)
     assert code_invoke == 404
-    assert 'Model {} is not loaded yet.'.format(model_name) in str(y2)
+    assert y2['error'].startswith('Servable not found for request')
+
+
+def test_list_models_empty():
+    code, res = make_list_model_request()
+    assert code == 200
+    assert res == {'models': []}
+
+
+def test_container_start_invocation_fail():
+    x = {
+        'instances': [1.0, 2.0, 5.0]
+    }
+    code, y = make_invocation_request(json.dumps(x), 'half_plus_three')
+    assert code == 404
+    assert y['error'].startswith('Servable not found for request')
 
 
 def test_load_two_models():
@@ -133,7 +162,7 @@ def test_load_two_models():
     }
     code1, res1 = make_load_model_request(json.dumps(model_data_1))
     assert code1 == 200
-    assert 'Successfully loaded model {}'.format(model_name_1) in res1
+    assert res1 == 'Successfully loaded model {}'.format(model_name_1)
 
     # load second model
     model_name_2 = 'half_plus_three'
@@ -143,26 +172,34 @@ def test_load_two_models():
     }
     code2, res2 = make_load_model_request(json.dumps(model_data_2))
     assert code2 == 200
-    assert 'Successfully loaded model {}'.format(model_name_2) in res2
+    assert res2 == 'Successfully loaded model {}'.format(model_name_2)
 
     # make invocation request to the first model
     x = {
         'instances': [1.0, 2.0, 5.0]
     }
     code_invoke1, y1 = make_invocation_request(json.dumps(x), model_name_1)
-    y1 = json.loads(y1)
     assert code_invoke1 == 200
     assert y1 == {'predictions': [2.5, 3.0, 4.5]}
 
     # make invocation request to the second model
     code_invoke2, y2 = make_invocation_request(json.dumps(x), 'half_plus_three')
-    y2 = json.loads(y2)
     assert code_invoke2 == 200
     assert y2 == {'predictions': [3.5, 4.0, 5.5]}
 
     code_list, res3 = make_list_model_request()
-    res3 = json.loads(res3)
-    assert len(res3) == 2
+    res3 = res3['models']
+    models = [json.loads(model) for model in res3]
+    assert code_list == 200
+    assert models == [
+        {
+            "modelName": "half_plus_two",
+            "modelUrl": "/opt/ml/models/half_plus_two"
+        },
+        {
+            "modelName": "half_plus_three",
+            "modelUrl": "/opt/ml/models/half_plus_three"
+        }]
 
 
 def test_load_one_model_two_times():
@@ -173,11 +210,11 @@ def test_load_one_model_two_times():
     }
     code_load, res = make_load_model_request(json.dumps(model_data))
     assert code_load == 200
-    assert 'Successfully loaded model {}'.format(model_name) in res
+    assert res == 'Successfully loaded model {}'.format(model_name)
 
     code_load2, res2 = make_load_model_request(json.dumps(model_data))
     assert code_load2 == 409
-    assert'Model {} is already loaded'.format(model_name) in res2
+    assert res2 == 'Illegal to list model {} multiple times in config list'.format(model_name)
 
 
 def test_load_non_existing_model():
@@ -189,7 +226,7 @@ def test_load_non_existing_model():
     }
     code, res = make_load_model_request(json.dumps(model_data))
     assert code == 404
-    assert 'Could not find valid base path {} for servable {}'.format(base_path, model_name) in str(res)
+    assert res == 'Could not find valid base path {} for servable {}'.format(base_path, model_name)
 
 
 def test_bad_model_reqeust():
@@ -210,4 +247,4 @@ def test_invalid_model_version():
     }
     code, res = make_load_model_request(json.dumps(invalid_model_version_data))
     assert code == 404
-    assert 'Could not find valid base path {} for servable {}'.format(base_path, model_name) in str(res)
+    assert res == 'Could not find valid base path {} for servable {}'.format(base_path, model_name)
